Eneable run.py that also runs events
